Update Load Image Batch
Load Image Batch update - WAS Suite now stores it's settings on it's own. Label specifies the ID of the node and thus it's counter for batching. If you change the directory, or pattern, the counter will reset.  The pattern is a glob pattern, so you can do things like "*.jpeg" to load only jpeg images, or "**/*" to load all images in the directory + sub directories.