Update WAS_Node_Suite.py
Masking is pretty much broken. I don't understand tensors well can't can't figure out how to get a good "mask to pil" and "pil to mask" system going that is compatible with ksampler etc and also able to be easily handleable with other operations or converting for preview image.

If someone knows how to fix this I'd love to hear.