Update WAS_Node_Suite.py
WAS_Image_Save
an images output was added to firstly force saving before the images are sent to the next group/node and secondly to have the possibility to bypass the node. so that saving can be switched on/off if necessary and the images can still be forwarded through the node to the next node in line.