Update WAS_Node_Suite.py
I like the save text node but noticed some issues that needed improvement such as no ability to add a suffix, the bug of no proper incremental files being saved with a delimiter. So if I don't use a delimiter the files are numbered after the total amount of files in the save folder. So if the folder has 124 images then the txt file saves as 125. A filter was added to ignore other files except .txt.

The code edits were required so that batch saving of .txt files can be done, with incremental files starting at 1, with the ability to add a prefix so that the match the image that caption was created from.